Enforce slave cleanup upon task completion when retention time is zero
Because the main slave cleanup mechanism is a periodic task, a retention
time of zero -which means 'no slave re-use'- isn't properly enforced:
slaves can still be re-used in between runs of the main slave cleanup
thread.

This change forces slave cleanup to happen after a task completes when
the retention time is set to zero to make sure that slave re-use can't
happen in this case.